requirements.txt
fastapi==0.115.3
motor==3.6.0
pymongo==4.9.2
python-dotenv==1.0.1
uvicorn==0.32.0
bcrypt==4.2.0
passlib==1.7.4
requests==2.32.3
pydantic==2.6.3
dnspython==2.4.2
typing-extensions==4.8.0